+++ .gitignore
@@ -175,7 +175,6 @@ dist/
 custom-elements.json
 **/src/react.ts
 **/stats.html
-**/component-statuses.json
 
 # Yalc
 .yalc
