+++ apps/pie-docs/package.json
@@ -12,7 +12,6 @@
     "build:dev": "DEPLOYMENT_ENVIRONMENT='development' eleventy",
     "clean": "run -T rimraf dist",
     "dev": "npx @11ty/eleventy --serve --port=8080",
-    "copy:component-statuses": "cp -v ../../component-statuses.json .",
     "lint:scripts": "run -T eslint .",
     "lint:scripts:fix": "run -T eslint . --fix",
     "lint:style": "run -T stylelint ./src/**/*.{css,scss}",
