+++ apps/pie-storybook/decorators/component-status.ts
@@ -1,7 +1,7 @@
 import { type TemplateResult, html } from 'lit';
 import '@justeattakeaway/pie-tag';
 import '@justeattakeaway/pie-icons-webc/dist/IconHelpCircle.js';
-import componentStatusesJson from '../component-statuses.json';
+import componentStatusesJson from '../../../component-statuses.json';
 import { type StorybookContext } from '../interfaces/storybook-context';
 import { type ComponentStatuses } from '../interfaces/component-status';
 import { tagVariantToStatusMap } from '../data/tag-variants-to-statuses-map';
