+++ apps/pie-storybook/turbo.json
@@ -8,21 +8,19 @@
       "cache": false,
       "dependsOn": [
         "^build",
-        "copy:component-statuses"
+        "//#generate:component-statuses"
       ]
     },
     "build": {
       "cache": true,
+      "inputs": [
+        "$TURBO_DEFAULT$",
+        "../../component-statuses.json"
+      ],
       "dependsOn": [
         "^build",
-        "copy:component-statuses"
-      ]
-    },
-    "copy:component-statuses": {
-      "dependsOn": [
         "//#generate:component-statuses"
-      ],
-      "cache": false
+      ]
     }
   }
 }
