+++ component-statuses.json
@@ -1,0 +1,24 @@
+{
+  "pie-assistive-text": "beta",
+  "pie-button": "beta",
+  "pie-card": "beta",
+  "pie-checkbox": "beta",
+  "pie-checkbox-group": "beta",
+  "pie-chip": "beta",
+  "pie-cookie-banner": "stable",
+  "pie-divider": "beta",
+  "pie-form-label": "alpha",
+  "pie-icon-button": "beta",
+  "pie-link": "beta",
+  "pie-lottie-player": "alpha",
+  "pie-modal": "beta",
+  "pie-notification": "beta",
+  "pie-radio": "alpha",
+  "pie-radio-group": "alpha",
+  "pie-spinner": "beta",
+  "pie-switch": "alpha",
+  "pie-tag": "beta",
+  "pie-text-input": "beta",
+  "pie-textarea": "alpha",
+  "pie-toast": "alpha"
+}
